update user entities
