+++ MusicWorld/Models/Entities/Album.cs
@@ -4,20 +4,20 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MusicWorld.Models
+namespace MusicWorld.Models.Entities
 {
-    public class Event
+    public class Album
     {
         public int Id { get; set; }
-        [Required]
         public string Title { get; set; }
         public string Description { get; set; }
-        public string TicketPrice { get; set; }
-        public string Place { get; set; }
-        public string StartingDate { get; set; }
+
         [Display(Name = "Image")]
         public string UrlImage { get; set; }
 
+        public string ReleaseDate { get; set; }
+
+        public ICollection<Song> Songs { get; set; }
         public ICollection<Artists> Artists { get; set; }
     }
 }
