+++ MusicWorld/Models/Entities/AppRole.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Identity;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace MusicWorld.Models.Entities
+{
+    public class AppRole : IdentityRole<Guid>
+    {
+        public string Description { get; set; }
+    }
+}
