+++ MusicWorld/Models/Entities/AppUser.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.AspNetCore.Identity;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace MusicWorld.Models.Entities
+{
+    public class AppUser : IdentityUser<Guid>
+    {
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+
+        public DateTime Dob { get; set; }
+    }
+}
