+++ MusicWorld/Models/Entities/Artists.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MusicWorld.Models
+namespace MusicWorld.Models.Entities
 {
     public class Artists
     {
@@ -13,12 +13,13 @@ namespace MusicWorld.Models
         public string Birthday { get; set; }
         public string Gender { get; set; }
         public string Contact { get; set; }
+
         [Display(Name = "Image")]
         public string UrlImage { get; set; }
 
         public int EventId { get; set; }
         public Event Events { get; set; }
-        
+
         public ICollection<Song> Songs { get; set; }
         public ICollection<Album> Albums { get; set; }
     }
