+++ MusicWorld/Models/Entities/Event.cs
@@ -4,18 +4,23 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MusicWorld.Models
+namespace MusicWorld.Models.Entities
 {
-    public class Album
+    public class Event
     {
         public int Id { get; set; }
+
+        [Required]
         public string Title { get; set; }
+
         public string Description { get; set; }
+        public string TicketPrice { get; set; }
+        public string Place { get; set; }
+        public string StartingDate { get; set; }
+
         [Display(Name = "Image")]
         public string UrlImage { get; set; }
-        public string ReleaseDate { get; set; }
 
-        public ICollection<Song> Songs { get; set; }
         public ICollection<Artists> Artists { get; set; }
     }
 }
