+++ MusicWorld/Models/IMusicReposity.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MusicWorld.Models.Entities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
