+++ MusicWorld/Models/MusicDbContext.cs
@@ -1,21 +1,22 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+using MusicWorld.Models.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MusicWorld.Models 
+namespace MusicWorld.Models
 {
-    public class MusicDbContext : DbContext
+    public class MusicDbContext : IdentityDbContext<AppUser, AppRole, Guid>
     {
         public MusicDbContext(DbContextOptions<MusicDbContext> options) : base(options)
         {
-
         }
+
         public DbSet<Event> Events { get; set; }
         public DbSet<Album> Albums { get; set; }
         public DbSet<Artists> Artists { get; set; }
         public DbSet<Song> Songs { get; set; }
-        public DbSet<User> Users { get; set; }
     }
 }
