--- MusicWorld/Models/User.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MusicWorld.Models
-{
-    public class User
-    {
-        public long Id { get; set; }
-        public string Name { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public string Email { get; set; }
-        public string Birthday { get; set; }
-        public string Gender { get; set; }
-        public string Address { get; set; }
-        public string PhoneNumber { get; set; }
-        public string UrlImage { get; set; }
-    }
-}
