+++ MusicWorld/Models/ViewModel/ProductListViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MusicWorld.Models.Entities;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -9,7 +10,7 @@ namespace MusicWorld.Models.ViewModel
     {
         public IEnumerable<Song> Songs { get; set; }
         public IEnumerable<Album> Albums { get; set; }
-        public IEnumerable<Event> Events{ get; set; }
+        public IEnumerable<Event> Events { get; set; }
         public IEnumerable<Artists> Artists { get; set; }
 
         public string CurrentCategory { get; set; }
